pyproject.toml
[tool.poetry]
name = "reflekt"
version = "0.3.0"
description = "CLI to design, collect, validate, stream, model, and document event data."
authors = ["Gregory Clunies <greg.clunies@gmail.com>"]
readme = "README.md"
include = ["reflekt/_templates/**/*"]

[tool.poetry.scripts]
reflekt = "reflekt.cli:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.6.1"
rich = "^12.5.1"
loguru = "^0.6.0"
PyYAML = "^6.0"
GitPython = "^3.1.27"
jsonschema = "^4.16.0"
python-dotenv = "^0.21.0"
inflection = "^0.5.1"
redshift-connector = "^2.0.909"
snowflake-sqlalchemy = "^1.4.4"

[tool.poetry.group.dev.dependencies]
autoflake = "^1.6.0"
black = "^22.8.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
make = "^0.1.6.post2"
mypy = "^0.971"
reuse = "^1.0.0"
pytest = "^7.1.3"
tox = "^3.26.0"
coverage = "^6.5.0"
pytest-cov = "^4.0.0"
flake8-docstrings = "^1.6.0"
darglint = "^1.8.1"
toml = "^0.10.2"
python-dotenv = "^0.21.0"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
known_first_party = 'reflekt'
multi_line_output = 3
lines_after_imports = 2
force_grid_wrap = 0
combine_as_imports = true
include_trailing_comma = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-vv -p no:warnings --cov=reflekt tests/"
testpaths = ["tests"]

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = true
envlist = py38,py39,py310

[testenv]
whitelist_externals = poetry
commands =
    poetry install -v
    poetry run pytest tests/

[flake8]
exclude =
  tests
"""

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
